+++ src/main/java/com/example/unbox_be/domain/trade/dto/request/SellingBidsPriceUpdateRequestDto.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class SellingBidsPriceUpdateRequestDto {
 
     @NotNull(message = "수정할 가격을 입력해주세요.")
     @Positive(message = "가격은 0보다 커야 합니다.")
-    private Integer newPrice;
+    private BigDecimal newPrice;
 }
